+++ src/lib.rs
@@ -7,20 +7,21 @@ Spelling correction & Fuzzy search based on Symmetric Delete spelling correction
 Single word spelling correction
 ```rust
 use symspell_rs::{SymSpell, Verbosity};
+use std::path::Path;
 
 let max_edit_distance_dictionary = 2; //maximum edit distance per dictionary precalculation
-let mut symspell: SymSpell = SymSpell::new(max_edit_distance_dictionary, 7, 1);
+let mut symspell: SymSpell = SymSpell::new(max_edit_distance_dictionary,None, 7, 1);
 
 // single term dictionary
-let term_index:i64 = 0; //column of the term in the dictionary text file
-let count_index:i64 = 1; //column of the term frequency in the dictionary text file
-symspell.load_dictionary("data/frequency_dictionary_en_82_765.txt", term_index, count_index, " ");
+let term_index = 0; //column of the term in the dictionary text file
+let count_index = 1; //column of the term frequency in the dictionary text file
+symspell.load_dictionary(Path::new("data/frequency_dictionary_en_82_765.txt"), term_index, count_index, " ");
 
 //lookup suggestions for single-word input strings
 let input_term = "hous";
 let suggestion_verbosity = Verbosity::Closest;//Top, Closest, All
 let max_edit_distance_lookup = 1; //max edit distance per lookup (maxEditDistanceLookup<=maxEditDistanceDictionary)
-let suggestions = symspell.lookup(input_term, suggestion_verbosity, max_edit_distance_lookup,false);
+let suggestions = symspell.lookup(input_term, suggestion_verbosity, max_edit_distance_lookup,None,None,false);
 //display suggestions, edit distance and term frequency
 println!("{:?}", suggestions);
 ```
@@ -28,16 +29,17 @@ println!("{:?}", suggestions);
 Compound aware multi-word spelling correction
 ```rust
 use symspell_rs::{SymSpell, Verbosity};
+use std::path::Path;
 
 let max_edit_distance_dictionary = 2; //maximum edit distance per dictionary precalculation
-let mut symspell = SymSpell::new(max_edit_distance_dictionary, 7, 1);
+let mut symspell = SymSpell::new(max_edit_distance_dictionary, None,7, 1);
 
 // single term dictionary
-let term_index:i64 = 0; //column of the term in the dictionary text file
-let count_index:i64 = 1; //column of the term frequency in the dictionary text file
-symspell.load_dictionary("data/frequency_dictionary_en_82_765.txt", term_index, count_index, " ");
+let term_index = 0; //column of the term in the dictionary text file
+let count_index = 1; //column of the term frequency in the dictionary text file
+symspell.load_dictionary(Path::new("data/frequency_dictionary_en_82_765.txt"), term_index, count_index, " ");
 // bigram dictionary
-symspell.load_bigram_dictionary("data/frequency_bigramdictionary_en_243_342.txt",0,2, " ",
+symspell.load_bigram_dictionary(Path::new("data/frequency_bigramdictionary_en_243_342.txt"),0,2, " ",
 );
 
 //lookup suggestions for multi-word input strings (supports compound splitting & merging)
@@ -51,14 +53,15 @@ println!("{:?}", compound_suggestions);
 Word Segmentation of noisy text
 ```rust
 use symspell_rs::{SymSpell, Verbosity};
+use std::path::Path;
 
 let max_edit_distance_dictionary = 0; //maximum edit distance per dictionary precalculation
-let mut symspell = SymSpell::new(max_edit_distance_dictionary, 7, 1);
+let mut symspell = SymSpell::new(max_edit_distance_dictionary, None,7, 1);
 
 // single term dictionary
-let term_index:i64 = 0; //column of the term in the dictionary text file
-let count_index:i64 = 1; //column of the term frequency in the dictionary text file
-symspell.load_dictionary("data/frequency_dictionary_en_82_765.txt", term_index, count_index, " ");
+let term_index = 0; //column of the term in the dictionary text file
+let count_index = 1; //column of the term frequency in the dictionary text file
+symspell.load_dictionary(Path::new("data/frequency_dictionary_en_82_765.txt"), term_index, count_index, " ");
 
 //word segmentation and correction for multi-word input strings with/without spaces
 let input_sentence = "thequickbrownfoxjumpsoverthelazydog";
@@ -71,14 +74,15 @@ println!("{:?}", result.segmented_string);
 Word Segmentation of Chinese text
 ```rust
 use symspell_rs::{SymSpell, Verbosity};
+use std::path::Path;
 
 let max_edit_distance_dictionary = 0; //maximum edit distance per dictionary precalculation
-let mut symspell = SymSpell::new(max_edit_distance_dictionary, 7, 1);
+let mut symspell = SymSpell::new(max_edit_distance_dictionary,None, 7, 1);
 
 // single term dictionary
-let term_index:i64 = 0; //column of the term in the dictionary text file
-let count_index:i64 = 1; //column of the term frequency in the dictionary text file
-symspell.load_dictionary("data/frequency_dictionary_zh_cn_349_045.txt", term_index, count_index, " ");
+let term_index = 0; //column of the term in the dictionary text file
+let count_index = 1; //column of the term frequency in the dictionary text file
+symspell.load_dictionary(Path::new("data/frequency_dictionary_zh_cn_349_045.txt"), term_index, count_index, " ");
 
 //word segmentation and correction for multi-word input strings with/without spaces
 let input_sentence = "部分居民生活水平";
@@ -92,4 +96,4 @@ println!("{:?}", result.segmented_string);
 
 mod symspell;
 mod test;
-pub use symspell::{Suggestion, SymSpell, Verbosity};
+pub use symspell::{Suggestion, SymSpell, Verbosity, damerau_levenshtein_osa};